added systemjs and typescript
Note that the current state is inconsistent; it only runs when loading order of
the different javascript files happens to be correct.

We now include a TypeScript to JavaScript compiler and SystemJS. This is the
starting point to loading all JavaScripts by importing them as modules, rather
than including them as <script> tags in index.html.

`npm run typings` will install necessary type definitions for AngularJS 1.x.
`npm run tsc` will watch for changes to .ts files and compile them to .js files.